% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMAD_bin_dataset.R
\name{IMAD_bin_dataset}
\alias{IMAD_bin_dataset}
\title{Generate usable binary data set from a}
\usage{
IMAD_bin_dataset(
  treatment,
  values,
  data,
  burnin = 10,
  pclin,
  seed = NA,
  trtlvls = c("A", "B"),
  weight = 0.5,
  alpha = 0.05,
  power = 0.8
)
}
\arguments{
\item{treatment}{A column containing the assigned treatments so far}

\item{values}{A column containing the values from the treatments}

\item{data}{A data set containing the above columns}

\item{burnin}{The number of burn-in values desired for each treatment group}

\item{pclin}{A vector that takes the minimum clinically significant success probabilities of each group}

\item{seed}{A set seed to help with reproducibility}

\item{trtlvls}{A vector containing the labels of each respective group}

\item{weight}{The weight of Sample Size Re-Estimation and Random Adaptive Re-Allocation}

\item{alpha}{The Type 1 error rate}

\item{power}{The desired power for the study}
}
\value{
A data set containing the previous data set with the attached next allocations.
}
\description{
Generate usable binary data set from a
}
\examples{
IMAD_bin_dataset
}
