% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMAD_next_val_bin.R
\name{IMAD_next_val_bin}
\alias{IMAD_next_val_bin}
\title{Allocation using IMAD for Binary Variables}
\usage{
IMAD_next_val_bin(
  group_a,
  group_b,
  trtlvls = c("A", "B"),
  p_clin,
  weight = 0.5,
  alpha = 0.05,
  beta = 0.2
)
}
\arguments{
\item{group_a}{A vector containing the values of the first group}

\item{group_b}{A vector containing the values of the second group}

\item{trtlvls}{A vector containing the labels of each respective group}

\item{p_clin}{A vector that takes the minimum clinically significant success probabilities of each group}

\item{weight}{The weight of Sample Size Re-Estimation and Random Adaptive Re-Allocation}

\item{alpha}{The Type 1 error rate}

\item{beta}{The Type 2 error rate}
}
\value{
The grouping of the next value and the expected sample sizes for each group given current estimated probabilities
}
\description{
Allocation using IMAD for Binary Variables
}
\examples{
IMAD_next_val_bin()
}
