% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMAD_simulation_cont.R
\name{IMAD_simulation_cont}
\alias{IMAD_simulation_cont}
\title{Simulation Code for IMAD Continuous Case}
\usage{
IMAD_simulation_cont(mu, sig, n, nu, weight = 0.5, alpha = 0.05, beta = 0.2)
}
\arguments{
\item{mu}{A vector giving the expected means of two groups}

\item{sig}{A vector giving the expected standard deviations of two groups}

\item{n}{A vector giving the initial counts given to each group}

\item{nu}{The preset clinically meaningful difference to be looked for}

\item{weight}{A weighting of Sample Size Re-estimation and Random Adaptive Re-Allocation}

\item{alpha}{The Type 1 error rate}

\item{beta}{The Type 2 error rate}
}
\value{
Gives the total number of patients to be allocated in each group as well as total sample size for that iteration of the simulation
}
\description{
Given a pair of preset means, standard deviations, and initial sample sizes along with weight and the minimally meaningful clinical difference,
a single simulation result of what a theoretical trial would look like along with
}
\examples{
IMAD_simulation_cont(mu = c(2, 0), sig = c(3, 3),  nu = 1.3, n = c(20, 20), weight = 0.3)
}
