% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_nought.R
\name{pi_nought}
\alias{pi_nought}
\title{Title}
\usage{
pi_nought(pa_clin = 0.6, pb_clin = 0.5, weight = 0.5, cme = 0.001)
}
\arguments{
\item{pa_clin}{The minimum clinically significant success probability of the first group}

\item{pb_clin}{The minimum clinically significant success probability of the second group}

\item{weight}{A weighting of Sample Size Re-estimation and Random Adaptive Re-Allocation}

\item{cme}{The set constraint to ensure the difference is at least as great as the clinically meaningful difference}
}
\value{
The optimal allocation ratio derived from the minimally significant difference in success probabilities between two groups
}
\description{
Title
}
\examples{
pi_nought()
}
