% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_binomial.R
\name{random_binomial}
\alias{random_binomial}
\title{Get Random Output from Two Binomial Distributions}
\usage{
random_binomial(n = c(20, 20), p = c(0.5, 0.5))
}
\arguments{
\item{n}{A vector containing the number of samples allocated to each group}

\item{p}{A vector containing the success probabilities of each group}
}
\value{
The binomial results from each group, the number of samples in each group, and the probabilities assigned to each group
}
\description{
Get Random Output from Two Binomial Distributions
}
\examples{
random_binomial()
random_binomial(n = c(10, 10), p = c(0.9, 0.1))
}
