% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_normal.R
\name{random_normal}
\alias{random_normal}
\title{Get Random Distribution from Two Normal Distributions}
\usage{
random_normal(n = c(20, 20), mu = c(-10, 0), sig = c(1, 1))
}
\arguments{
\item{n}{A vector containing the number of samples allocated to each group}

\item{mu}{A vector of means of the two normal distributions}

\item{sig}{A vector of standard deviations of the two normal distributions}
}
\value{
Results from the random distributions along with the preset values given
}
\description{
Get Random Distribution from Two Normal Distributions
}
\examples{
random_normal()
random_normal(n = c(10, 10), mu = c(-1, 0), sig = c(2, 2))
}
