% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_objective_bin.R
\name{weighted_objective_bin}
\alias{weighted_objective_bin}
\title{Weighted Objective Functions for Binary Outcomes}
\usage{
weighted_objective_bin(
  pa_clin = 0.6,
  pb_clin = 0.5,
  pa_hat = 0.6,
  pb_hat = 0.5,
  fa = 0.5,
  fb = 0.5,
  weight = 0.5,
  n_a = 10,
  n_b = 10,
  n = 20,
  zalpha = qnorm(0.05/2),
  zbeta = qnorm(0.2)
)
}
\arguments{
\item{pa_clin}{The minimum clinically significant success probability of the first group}

\item{pb_clin}{The minimum clinically significant success probability of the second group}

\item{pa_hat}{The current expected probability of the first group}

\item{pb_hat}{The current expected probability of the second group}

\item{fa}{Location invariant value of the first group}

\item{fb}{Location invariant value of the second group}

\item{weight}{The weight of Sample Size Re-Estimation and Random Adaptive Re-Allocation}

\item{n_a}{The number of individuals currently in the first group}

\item{n_b}{The number of individuals currently in the second group}

\item{n}{The total number of individuals currently in either group}

\item{zalpha}{The standardized set type 1 error rate}

\item{zbeta}{The standardized set type 2 error rate}
}
\value{
Returns optimal allocation, optimal sample size ratio, and allocation of the next patient for binary outcomes
}
\description{
Weighted Objective Functions for Binary Outcomes
}
\examples{
weighted_objective_bin()
}
